fix: improvements from self-review
