+++ libs/explorers/util/src/index.ts
@@ -7,7 +7,6 @@ export * from './lib/modal-link/modal-link.component';
 export * from './lib/pipes';
 export * from './lib/popover-link/popover-link.component';
 export * from './lib/prefetch-config';
-export * from './lib/string/capitalize-first-letter';
 export * from './lib/svg-icon/svg-icon.component';
 export * from './lib/tooltip-button/tooltip-button.component';
 export * from './lib/wiki/wiki.component';
