--- libs/explorers/util/src/lib/string/capitalize-first-letter.ts
@@ -1,4 +1,0 @@
-export function capitalizeFirstLetter(value: string | null | undefined): string {
-  if (!value) return '';
-  return value.charAt(0).toUpperCase() + value.slice(1);
-}
